Start reorganizing directories to chapter/type
I started off organizing the exercises by type/chapter, and for
projects I didn't do a good job of organizing them at all. This
is a step towards reorginizing the directories to be primarily
by chapter and then secondarily by exercise type.

So instead of having:
    pytest-exercises/3-functions/...

You would have
    3-functions/pytest-exercises/...

I'm doing this because I want all the content for a unit to be
bundled together. Hopefully it will make it easier for the
students to navigate, and hopefully it will make it easier for
us to see what content has weak coverage.

This commit puts all the pytest-exercises and partial-projects
in the right place. Before I can put the whole projects in the
right place, I need to look at them more closely to figure out
where they belong.